test/helpers/pouch: Destroy only existing DB
  A lot of tests use the `cleanDatase()` PouchDB test helper which
  destroys any existing database.
  However, it seems that this helper can be called multiple times and
  willl fail on any second call as the DB to destroy does not exist
  anymore.

  As a first mitigation step, we make sure the helper tries to destroy
  the PouchDB database only if it exists.
  A second step would consist in finding and removing those duplicate
  calls.

+++ test/support/helpers/pouch.js
@@ -7,7 +7,9 @@ module.exports = {
   },
 
   async cleanDatabase() {
-    await this.pouch.db.destroy()
+    if (this.pouch && this.pouch.db) {
+      await this.pouch.db.destroy()
+    }
     this.pouch = null
   },
